requirements.txt
Flask
pandas
openpyxl
groq
pandasql
python-dotenv
matplotlib